requirements.txt
pycryptodome==3.4.7
flask=0.12.2